Create spawn method refactored
Refactored the create spawn method and added more abstraction levels to it. Cleaned up the code, and made it easier to change how large the diameter of the spawn should be by adding a global variable to the class. This made it easier to set the "middle point" for the enchantment table, and where the wool should switch colors.